+++ src/components/Course/Form/index.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback } from 'react';
+import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   Autocomplete,
@@ -13,12 +13,8 @@ import {
   TextField,
   styled
 } from '@mui/material';
-import * as Yup from 'yup';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { useForm, Controller } from 'react-hook-form';
-import request from 'src/hooks/useRequest';
-import { Category } from 'src/models/category';
-import { useNavigate } from 'react-router-dom';
+import { Controller } from 'react-hook-form';
+import { useCourseForm } from './store';
 
 const FormLabelStyle = styled('p')(
   () => `
@@ -27,73 +23,19 @@ const FormLabelStyle = styled('p')(
   `
 );
 
-type FormInputType = {
-  courseName: string;
-  categoryId: number;
-  name: string;
-  price: number;
-};
-
 const Form = () => {
-  const navigate = useNavigate();
-  const [categories, setCategories] = useState<Category[]>([]);
-  const { t }: { t: any } = useTranslation();
-
-  const schema = Yup.object({
-    categoryId: Yup.number().required(t('Need select category.')),
-    name: Yup.string().required(t('Product name is required.')),
-    price: Yup.number()
-      .min(1, t('Must be a number greater than or equal to 1.'))
-      .required(t('Price is required.'))
-  }).required();
   const {
     register,
     control,
     handleSubmit,
     setValue,
-    formState: { errors, isSubmitting }
-  } = useForm<FormInputType>({
-    defaultValues: {
-      courseName: '',
-      categoryId: 0,
-      name: '',
-      price: 0
-    },
-    resolver: yupResolver(schema)
-  });
-  const onSubmit = (data: FormInputType) => {
-    try {
-      request({
-        url: '/v1/products',
-        method: 'POST',
-        reqParams: {
-          data: {
-            productType: 'コース',
-            categoryId: data.categoryId,
-            name: data.name,
-            price: data.price
-          }
-        }
-      }).then(() => {
-        navigate('/dashboards/courses');
-      });
-    } catch (e) {
-      console.error(e);
-    }
-  };
+    formState: { errors, isSubmitting },
+    onSubmit,
+    categories,
 
-  const getCategories = useCallback(() => {
-    try {
-      request({
-        url: '/v1/categories',
-        method: 'GET'
-      }).then((response) => {
-        setCategories(response.data.categories);
-      });
-    } catch (e) {
-      console.error(e);
-    }
-  }, []);
+    getCategories
+  } = useCourseForm();
+  const { t }: { t: any } = useTranslation();
 
   useEffect(() => {
     getCategories();
